+++ tests/Microplate/Scalars/Row8Test.php
@@ -5,10 +5,10 @@ namespace MLL\Utils\Tests\Microplate\Scalars;
 use GraphQL\Error\Error;
 use GraphQL\Error\InvariantViolation;
 use GraphQL\Type\Definition\ScalarType;
-use MLL\Utils\Microplate\Scalars\Row96Well;
+use MLL\Utils\Microplate\Scalars\Row8;
 use PHPUnit\Framework\TestCase;
 
-final class Row96WellTest extends TestCase
+final class Row8Test extends TestCase
 {
     protected function setUp(): void
     {
@@ -19,50 +19,50 @@ final class Row96WellTest extends TestCase
         }
     }
 
-    public function testSerializeThrowsIfRow96WellIsInvalid(): void
+    public function testSerializeThrowsIfInvalid(): void
     {
         $this->expectException(InvariantViolation::class);
         $this->expectExceptionMessage('The given value "I" did not match the regex /^[A-H]$/.');
 
-        (new Row96Well())->serialize('I');
+        (new Row8())->serialize('I');
     }
 
-    public function testSerializeThrowsIfRow96WellIsNonCapital(): void
+    public function testSerializeThrowsIfNonCapital(): void
     {
         $this->expectException(InvariantViolation::class);
         $this->expectExceptionMessage('The given value "h" did not match the regex /^[A-H]$/.');
 
-        (new Row96Well())->serialize('h');
+        (new Row8())->serialize('h');
     }
 
-    public function testSerializePassesWhenRow96WellIsValid(): void
+    public function testSerializePassesWhenValid(): void
     {
-        $serializedResult = (new Row96Well())->serialize('H');
+        $serializedResult = (new Row8())->serialize('H');
 
         self::assertSame('H', $serializedResult);
     }
 
-    public function testParseValueThrowsIfRow96WellIsInvalid(): void
+    public function testParseValueThrowsIfInvalid(): void
     {
         $this->expectException(Error::class);
         $this->expectExceptionMessage('The given value "I" did not match the regex /^[A-H]$/.');
 
-        (new Row96Well())->parseValue('I');
+        (new Row8())->parseValue('I');
     }
 
-    public function testParseValueThrowsIfRow96WellIsNonCapital(): void
+    public function testParseValueThrowsIfNonCapital(): void
     {
         $this->expectException(Error::class);
         $this->expectExceptionMessage('The given value "h" did not match the regex /^[A-H]$/.');
 
-        (new Row96Well())->parseValue('h');
+        (new Row8())->parseValue('h');
     }
 
-    public function testParseValuePassesIfRow96WellIsValid(): void
+    public function testParseValuePassesIfValid(): void
     {
         self::assertSame(
             'H',
-            (new Row96Well())->parseValue('H')
+            (new Row8())->parseValue('H')
         );
     }
 }